refactor(reviewer): extract fix-loop owner
