+++ src/main/reviewer/orchestrator.ts
@@ -7,21 +7,15 @@
 //
 // Errors are isolated: reviewer failures set lifecycle='error' and do NOT crash the main session.
 
-import { randomUUID } from 'node:crypto'
-
 import { withReviewerRuntimeActivity, type ReviewerAcpRuntime } from './acp-runtime'
 import { createLogger } from '../logger'
-import type { ReviewCheck, ReviewWithChecks } from '../../shared/reviewer'
+import type { ReviewWithChecks } from '../../shared/reviewer'
 import type { ReviewRepository } from './repository'
-import {
-  materializeSessionConversationGraph,
-  type PersistedChatSession
-} from '../../shared/session-persistence'
+import type { PersistedChatSession } from '../../shared/session-persistence'
 import type { ArtifactVersionContentResolver } from './host-sdk'
-import { injectAuditorMessage } from './correction'
 import { buildHistoryPreamble } from '../../shared/history-preamble'
-import { getActiveConversationContext } from '../../shared/conversation-graph'
-import { runReviewAssessment, type ReviewAssessmentResult } from './review-assessment-owner'
+import { runReviewAssessment } from './review-assessment-owner'
+import { runReviewerFixLoop } from './reviewer-fix-loop-owner'
 
 export { buildReviewerPrompt } from './review-assessment-owner'
 export { driveReviewerToStop } from './reviewer-session-driver'
@@ -114,455 +108,6 @@ export type RunReviewOptions = {
 const DEFAULT_REVIEWER_TIMEOUT_MS = 900_000
 const DEFAULT_REVIEWER_MAX_UPDATES = 1000
 const DEFAULT_SESSION_REFRESH_TIMEOUT_MS = 10_000
-const SESSION_REFRESH_POLL_MS = 50
-
-// Options for the Phase 3 fix loop.
-type FixLoopOptions = {
-  sessionId: string
-  // The original turn's message id (shared across all Review rows in this closure).
-  originalTurnMessageId: string
-  // The currently-open warn/fail checks to carry forward into each re-review.
-  openChecks: ReviewCheck[]
-  projectId: string
-  mainSessionId: string
-  getSession: SessionProvider
-  reviewRepository: ReviewRepository
-  runSessionMutation?: ReviewMutationRunner
-  acpRuntime: ReviewerAcpRuntime
-  artifactStorageRoot: string
-  artifactVersionContentResolver?: ArtifactVersionContentResolver
-  reviewerMcpEntryPath?: string
-  model: string
-  onReviewUpdate?: (review: ReviewWithChecks) => void
-  onCorrectionPrompt?: (text: string) => void
-  onCorrectionFailed?: () => void
-  reviewerTimeoutMs: number
-  reviewerMaxUpdates: number
-  maxRounds: number
-  sessionRefreshTimeoutMs: number
-  // Optional abort signal: when aborted, the loop exits at the next round boundary.
-  abortSignal?: AbortSignal
-}
-
-const waitForCorrectionAgentMessage = async (options: {
-  sessionId: string
-  messageIdsBefore: ReadonlySet<string>
-  getSession: SessionProvider
-  timeoutMs: number
-  abortSignal?: AbortSignal
-}): Promise<
-  | {
-      session: PersistedChatSession
-      message: PersistedChatSession['messages'][number]
-    }
-  | undefined
-> => {
-  const deadline = Date.now() + options.timeoutMs
-
-  for (;;) {
-    if (options.abortSignal?.aborted) return undefined
-
-    const latest = await options.getSession(options.sessionId)
-    const correction = latest?.messages.find(
-      (message) =>
-        !options.messageIdsBefore.has(message.id) &&
-        message.role === 'agent' &&
-        message.status === 'complete'
-    )
-    if (latest && correction) return { session: latest, message: correction }
-
-    if (Date.now() >= deadline) return undefined
-    await new Promise<void>((resolve) => setTimeout(resolve, SESSION_REFRESH_POLL_MS))
-  }
-}
-
-// Runs the Phase 3 bounded re-review loop. For each round (up to maxRounds):
-// 1. Injects [Auditor] with the still-open warn/fail checks.
-// 2. The main agent produces a correction turn.
-// 3. Re-reviews the correction turn's new blocks.
-// 4. Updates each original finding by its stable sourceFindingId:
-//    - pass → resolved
-//    - warn/fail → incrementReflagCount; stays open
-//    - missing/unknown/duplicate id → submission rejected, original stays open
-// 5. If all resolved or cap reached, stops.
-// Cap termination marks remaining open warn/fail checks as 'unaddressed'.
-const runFixLoop = async (options: FixLoopOptions): Promise<void> => {
-  const {
-    sessionId,
-    originalTurnMessageId,
-    projectId,
-    mainSessionId,
-    getSession,
-    reviewRepository,
-    runSessionMutation,
-    acpRuntime,
-    artifactStorageRoot,
-    artifactVersionContentResolver,
-    reviewerMcpEntryPath,
-    model,
-    onReviewUpdate,
-    onCorrectionPrompt,
-    onCorrectionFailed,
-    reviewerTimeoutMs,
-    reviewerMaxUpdates,
-    maxRounds,
-    sessionRefreshTimeoutMs,
-    abortSignal
-  } = options
-
-  let openChecks = [...options.openChecks]
-  const commitDispositionBatch = async (
-    inputs: Parameters<ReviewRepository['commitFindingDispositions']>[0]
-  ): Promise<void> => {
-    if (inputs.length === 0) return
-    await runReviewMutation(runSessionMutation, () =>
-      reviewRepository.commitFindingDispositions(inputs)
-    )
-
-    // Finding/disposition writes change the original Review card without creating a new Review row.
-    // Push each mutated row after commit so open Reviewer and Provenance surfaces reload immediately.
-    const mutatedReviewIds = new Set(inputs.map((input) => input.reviewId))
-    const reviews = await reviewRepository.getReviewsForProjectSession(projectId, sessionId)
-    for (const review of reviews) {
-      if (mutatedReviewIds.has(review.id)) onReviewUpdate?.(review)
-    }
-  }
-  const markOpenChecksUnaddressed = async (
-    trigger: 'loop_terminated' | 'correction_failed' | 'aborted',
-    note: string
-  ): Promise<void> => {
-    await commitDispositionBatch(
-      openChecks.map((openCheck) => ({
-        reviewId: openCheck.reviewId,
-        sourceFindingId: openCheck.id,
-        trigger,
-        outcome: 'unaddressed',
-        note,
-        assessedArtifactVersionId: openCheck.artifactVersionId
-      }))
-    )
-  }
-
-  for (let round = 0; round < maxRounds; round++) {
-    if (openChecks.length === 0) break
-
-    // Abort check: if the user cancelled during the loop, exit without further [Auditor] injections.
-    if (abortSignal?.aborted) {
-      log.info('fix loop: aborted by user', { sessionId, round, openCount: openChecks.length })
-      await markOpenChecksUnaddressed('aborted', 'The fix loop was aborted by the user.')
-      return
-    }
-
-    // Step A: record every known message id before the correction prompt. The provider is awaited on
-    // every use; production reloads durable storage rather than returning the initial review snapshot.
-    let sessionBefore: PersistedChatSession | undefined
-    try {
-      sessionBefore = await getSession(sessionId)
-    } catch (error) {
-      log.warn('fix loop: failed to load durable session before correction', {
-        sessionId,
-        round,
-        error: error instanceof Error ? error.message : String(error)
-      })
-      await markOpenChecksUnaddressed('correction_failed', 'Could not load the durable session.')
-      return
-    }
-    if (!sessionBefore) {
-      log.warn('fix loop: durable session disappeared before correction', { sessionId, round })
-      await markOpenChecksUnaddressed('correction_failed', 'The durable session disappeared.')
-      return
-    }
-    const messagesBefore = sessionBefore.messages
-    const messageIdsBefore = new Set(messagesBefore.map((message) => message.id))
-
-    // Step B: inject [Auditor] with the currently-open warn/fail checks.
-    let correctionFailed = false
-    try {
-      const provenanceContext = getActiveConversationContext(
-        materializeSessionConversationGraph(sessionBefore).conversationGraph!,
-        `prompt-${randomUUID()}`
-      )
-      await injectAuditorMessage({
-        sessionId,
-        mainSessionId,
-        findings: openChecks,
-        acpRuntime,
-        provenanceContext,
-        onCorrectionPrompt,
-        onCorrectionFailed: () => {
-          correctionFailed = true
-          onCorrectionFailed?.()
-        }
-      })
-    } catch (error) {
-      correctionFailed = true
-      log.warn('fix loop: failed to derive correction provenance', {
-        sessionId,
-        round,
-        error: error instanceof Error ? error.message : String(error)
-      })
-    }
-
-    // Error handling: a failed correction counts as a round (prevents infinite loop) but we
-    // cannot re-review (there's no correction turn). Mark remaining as unaddressed and stop.
-    if (correctionFailed) {
-      log.warn('correction failed in fix loop — marking remaining checks unaddressed', {
-        sessionId,
-        round,
-        openCount: openChecks.length
-      })
-      await markOpenChecksUnaddressed('correction_failed', 'The correction prompt failed.')
-      return
-    }
-
-    // Step C: wait for the new agent message to reach durable storage. sendPrompt completion and the
-    // renderer persistence queue are independent, so an immediate one-shot reload is still racy.
-    let correctionState:
-      | { session: PersistedChatSession; message: PersistedChatSession['messages'][number] }
-      | undefined
-    try {
-      correctionState = await waitForCorrectionAgentMessage({
-        sessionId,
-        messageIdsBefore,
-        getSession,
-        timeoutMs: sessionRefreshTimeoutMs,
-        abortSignal
-      })
-    } catch (error) {
-      log.warn('fix loop: failed while refreshing durable correction turn', {
-        sessionId,
-        round,
-        error: error instanceof Error ? error.message : String(error)
-      })
-      await markOpenChecksUnaddressed(
-        'correction_failed',
-        'Could not reload the durable correction turn.'
-      )
-      return
-    }
-    if (!correctionState) {
-      if (abortSignal?.aborted) {
-        log.info('fix loop: aborted while waiting for durable correction turn', {
-          sessionId,
-          round
-        })
-        await markOpenChecksUnaddressed('aborted', 'The fix loop was aborted by the user.')
-        return
-      }
-      log.warn('correction turn did not reach durable session storage; refusing stale re-review', {
-        sessionId,
-        round,
-        timeoutMs: sessionRefreshTimeoutMs
-      })
-      await markOpenChecksUnaddressed(
-        'correction_failed',
-        'The correction turn did not reach durable storage.'
-      )
-      return
-    }
-
-    const correctionTurnMessageId = correctionState.message.id
-
-    // Step D: run a re-review scoped to the correction turn's new blocks.
-    // This creates a new Review row sharing the original turnMessageId.
-    log.info('fix loop: running re-review', { sessionId, round, correctionTurnMessageId })
-
-    const scopedResult = await runScopedReview({
-      sessionId,
-      turnMessageId: correctionTurnMessageId,
-      originalTurnMessageId,
-      projectId,
-      getSession,
-      reviewRepository,
-      runSessionMutation,
-      acpRuntime,
-      artifactStorageRoot,
-      artifactVersionContentResolver,
-      reviewerMcpEntryPath,
-      model,
-      onReviewUpdate,
-      reviewerTimeoutMs,
-      reviewerMaxUpdates,
-      trackedChecks: openChecks,
-      sessionSnapshot: correctionState.session
-    })
-    const reReviewResult = scopedResult.review
-
-    // Step E: compute resolution transitions for the original review's open checks.
-    // - If the re-review errored: count as a round but mark remaining unaddressed and stop.
-    // - Each original finding is matched only by sourceFindingId, never by model-generated prose.
-    // - A pass disposition resolves it; warn/fail increments reflagCount and keeps it open.
-    // - Missing dispositions are rejected by MCP and stay open defensively if one slips through.
-
-    if (reReviewResult.lifecycle === 'error') {
-      log.warn('fix loop: re-review errored — marking remaining checks unaddressed', {
-        sessionId,
-        round,
-        openCount: openChecks.length
-      })
-      await markOpenChecksUnaddressed('correction_failed', 'The scoped re-review failed.')
-      return
-    }
-
-    const dispositionsByFindingId = new Map(
-      scopedResult.submittedChecks.flatMap((check) =>
-        check.sourceFindingId ? [[check.sourceFindingId, check] as const] : []
-      )
-    )
-
-    const stillOpenChecks: ReviewCheck[] = []
-
-    for (const openCheck of openChecks) {
-      const disposition = dispositionsByFindingId.get(openCheck.id)
-      if (!disposition) {
-        // The MCP server rejects incomplete submissions, so this is defensive fail-closed behavior.
-        log.error('fix loop: scoped re-review omitted a tracked finding disposition', {
-          sessionId,
-          round,
-          findingId: openCheck.id
-        })
-        stillOpenChecks.push(openCheck)
-      } else if (disposition.status === 'warn' || disposition.status === 'fail') {
-        log.info('fix loop: finding re-flagged', {
-          sessionId,
-          round,
-          findingId: openCheck.id
-        })
-        stillOpenChecks.push(openCheck)
-      } else {
-        log.info('fix loop: finding resolved', { sessionId, round, findingId: openCheck.id })
-      }
-    }
-
-    const newIssueSortIndexes = new Set(
-      scopedResult.submittedChecks
-        .filter(
-          (check) => !check.sourceFindingId && (check.status === 'warn' || check.status === 'fail')
-        )
-        .map((check) => check.sortIndex)
-    )
-    const newlyOpenChecks = reReviewResult.checks.filter(
-      (check) =>
-        newIssueSortIndexes.has(check.sortIndex) &&
-        (check.status === 'warn' || check.status === 'fail')
-    )
-    if (newlyOpenChecks.length > 0) {
-      log.info('fix loop: carrying newly discovered findings into the next round', {
-        sessionId,
-        round,
-        count: newlyOpenChecks.length
-      })
-    }
-
-    openChecks = [...stillOpenChecks, ...newlyOpenChecks]
-
-    if (openChecks.length === 0) {
-      log.info('fix loop: all checks resolved', { sessionId, rounds: round + 1 })
-      return
-    }
-
-    log.info('fix loop: still-open checks remain', {
-      sessionId,
-      round,
-      stillOpen: openChecks.length
-    })
-  }
-
-  // Cap reached: mark remaining open warn/fail checks as unaddressed.
-  if (openChecks.length > 0) {
-    log.info('fix loop: cap reached — marking remaining checks unaddressed', {
-      sessionId,
-      maxRounds,
-      remaining: openChecks.length
-    })
-    await markOpenChecksUnaddressed(
-      'loop_terminated',
-      `Fix loop reached its ${maxRounds}-round cap.`
-    )
-  }
-}
-
-// Runs one scoped re-review for a correction turn. Creates a new Review row under the same
-// original turnMessageId so all iterations are grouped. Returns the completed review.
-// Never throws — errors are captured as lifecycle='error'.
-const runScopedReview = async (options: {
-  sessionId: string
-  turnMessageId: string
-  originalTurnMessageId: string
-  projectId: string
-  getSession: SessionProvider
-  reviewRepository: ReviewRepository
-  runSessionMutation?: ReviewMutationRunner
-  acpRuntime: ReviewerAcpRuntime
-  artifactStorageRoot: string
-  artifactVersionContentResolver?: ArtifactVersionContentResolver
-  reviewerMcpEntryPath?: string
-  model: string
-  onReviewUpdate?: (review: ReviewWithChecks) => void
-  reviewerTimeoutMs: number
-  reviewerMaxUpdates: number
-  trackedChecks: ReviewCheck[]
-  sessionSnapshot?: PersistedChatSession
-}): Promise<ReviewAssessmentResult> => {
-  const {
-    sessionId,
-    turnMessageId,
-    originalTurnMessageId,
-    projectId,
-    getSession,
-    reviewRepository,
-    runSessionMutation,
-    acpRuntime,
-    artifactStorageRoot,
-    artifactVersionContentResolver,
-    reviewerMcpEntryPath,
-    model,
-    onReviewUpdate,
-    reviewerTimeoutMs,
-    reviewerMaxUpdates,
-    trackedChecks,
-    sessionSnapshot
-  } = options
-
-  // Keep missing-session handling in the facade: the assessment owner always receives a durable
-  // snapshot and therefore never invents lifecycle rows for absent sessions.
-  const session = sessionSnapshot ?? (await getSession(sessionId))
-  if (!session) {
-    log.warn('session not found for scoped re-review', { sessionId })
-    const errorReview = await runReviewMutation(runSessionMutation, () =>
-      reviewRepository.createReview({
-        projectId,
-        sessionId,
-        turnMessageId: originalTurnMessageId,
-        scope: { turnMessageId: originalTurnMessageId, blocks: [], artifactVersionIds: [] },
-        lifecycle: 'error',
-        errorMessage: `Session ${sessionId} not found during re-review`,
-        model
-      })
-    )
-    return { review: { ...errorReview, checks: [] }, submittedChecks: [] }
-  }
-
-  return runReviewAssessment({
-    mode: 'tracked',
-    session,
-    sessionId,
-    scopeTurnMessageId: turnMessageId,
-    turnMessageId: originalTurnMessageId,
-    projectId,
-    reviewRepository,
-    runSessionMutation,
-    acpRuntime,
-    artifactStorageRoot,
-    artifactVersionContentResolver,
-    reviewerMcpEntryPath,
-    model,
-    onReviewUpdate,
-    reviewerTimeoutMs,
-    reviewerMaxUpdates,
-    trackedChecks
-  })
-}
 
 // Drives one complete auto-review cycle and returns the final review (with checks) for the caller
 // to broadcast. Never throws — errors are captured as lifecycle='error'.
@@ -626,7 +171,7 @@ const runReviewWithSession = async (
   if (mainSessionId && hasWarnOrFail) {
     onFixLoopStart?.()
     try {
-      await runFixLoop({
+      await runReviewerFixLoop({
         sessionId,
         originalTurnMessageId: turnMessageId,
         openChecks: finalReview.checks.filter((c) => c.status === 'warn' || c.status === 'fail'),
@@ -655,7 +200,7 @@ const runReviewWithSession = async (
 
     // Reload checks after the fix loop so the returned object reflects final resolutions.
     const reloadedReviews = await reviewRepository.getReviewsForProjectSession(projectId, sessionId)
-    const reloadedReview = reloadedReviews.find((r) => r.id === finalReview.id)
+    const reloadedReview = reloadedReviews.find((review) => review.id === finalReview.id)
     if (reloadedReview) {
       onReviewUpdate?.(reloadedReview)
       return reloadedReview
